changing the types in a struct

+++ 0x0E-structures_typedef/dog.h
@@ -13,7 +13,7 @@
 struct dog
 {
 	char *name;
-	double age;
+	float age;
 	char *owner;
 };
 #endif /* DOG_H */
